config update
Give up on outer walls first. Bad layer starts that I can't fix
(inconsistent extrusion that fucks the whole surface until the next
corner). Optimise for inner walls first instead.

- reduce pressure advance and increase square corner velocity to
  combat bulging
  - scv at 10 reduces bulging with no discernible effect on corners,
    going over has diminishing returns and starts to round corners
- limit speeds to 100
- drop overhang speeds (still not happy with overhang performance
  but I'm asking for too much at this point)

Touch wood but in all this stuffing about the surface artifacts that
usually present with inner walls first are gone and the layer lines
look good. I'm gonna sneeze in the same room as the printer and it's
probably going to go to shit.